# docs/swagger/swagger.yaml
basePath: /api/v1
definitions:
  domain.DeptRole:
    enum:
    - preparator
    - reviewer
    - approver
    type: string
    x-enum-varnames:
    - DeptPreparator
    - DeptReviewer
    - DeptApprover
  domain.InvitationDeptMapping:
    properties:
      department_id:
        type: string
      level:
        $ref: '#/definitions/domain.DeptRole'
    type: object
...
  http.DeptMemberView:
    properties:
      department_id:
        type: string
      level:
        type: string
    type: object
...
  http.ErrorResponse:
    properties:
      active_workflows:
        description: |-
          ActiveWorkflows is populated on 409 workflow_resolution_required
          responses (§8.8). Count of open workflows blocking the mutation.
        example: 3
        type: integer
      allowed_actions:
        description: AllowedActions lists the P-26 resolution actions the caller may
          invoke.
        example:
        - replace_delegate
        - stop_workflows
        items:
          type: string
        type: array
      code:
        description: Code is a legacy alias for Error kept for backwards compatibility.
        example: not_found
        type: string
      details:
        description: Details is populated with per-field errors for 422 validation
          failures.
        items:
          $ref: '#/definitions/http.ValidationError'
        type: array
      error:
        description: Error is the machine-readable error code (LLD §17 error taxonomy).
        example: not_found
        type: string
      message:
        description: Message is a human-readable explanation of the error.
        example: resource not found
        type: string
      record_version:
        description: |-
          RecordVersion is populated on 409 optimistic_lock_conflict responses
          so the client can re-read and retry (CONC-4).
        example: 5
        type: integer
      request_id:
        description: RequestID is the per-request UUID injected by platform-gincommon.
        example: 01H8XYZ0000000000000000000
        type: string
      status:
        description: Status is the HTTP status code echoed in the body.
        example: 404
        type: integer
      trace_id:
        description: TraceID is the OTel W3C trace ID for cross-service correlation.
        example: a3f1b2c8d4e5f60718293a4b5c6d7e8f
        type: string
      workflow_ids:
        description: WorkflowIDs is populated on 409 workflow_resolution_required
          responses.
        items:
          type: string
        type: array
    type: object
...
  http.InvitationCreateRequest:
    properties:
      email:
        type: string
      full_name:
        type: string
      initial_dept_mappings:
        items:
          $ref: '#/definitions/domain.InvitationDeptMapping'
        type: array
      initial_tenant_roles:
        items:
          type: string
        type: array
    type: object
...
  http.InvitationResponse:
    properties:
      email:
        type: string
      expires_at:
        type: string
      full_name:
        type: string
      id:
        type: string
      record_version:
        type: integer
      status:
        type: string
    type: object
  http.MemberItemResponse:
    properties:
      departments:
        items:
          $ref: '#/definitions/http.DeptMemberView'
        type: array
      record_version:
        type: integer
      status:
        type: string
      tenant_roles:
        items:
          type: string
        type: array
      updated_at:
        type: string
      user_id:
        type: string
    type: object
  http.MembershipListResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/http.MemberItemResponse'
        type: array
      next_cursor:
        type: string
    type: object
  http.MembershipPatchRequest:
    properties:
      record_version:
        type: integer
      status:
        type: string
    type: object
  http.MembershipPatchResponse:
    properties:
      record_version:
        example: 2
        type: integer
      status:
        enum:
        - active
        - suspended
        type: string
      updated_at:
        format: date-time
        type: string
      user_id:
        format: uuid
        type: string
    type: object
...
  http.RolesPutRequest:
    properties:
      roles:
        items:
          type: string
        type: array
    type: object
  http.RolesReconcileResponse:
    properties:
      granted:
        example:
        - tenant_admin
        items:
          type: string
        type: array
      revoked:
        example:
        - tender_admin
        items:
          type: string
        type: array
    type: object
  http.SeatUsageResponse:
    properties:
      active_users:
        example: 12
        type: integer
      grace_ends_at:
        format: date-time
        type: string
      licensed_seats:
        example: 25
        type: integer
      over_cap:
        example: false
        type: boolean
      overage_since:
        format: date-time
        type: string
      pending_invitations:
        example: 3
        type: integer
    type: object
...
  http.ValidationError:
    properties:
      code:
        description: Code is the machine-readable violation code.
        example: out_of_range
        type: string
      field:
        description: Field is the JSON path of the failing field.
        example: mfa_freshness_seconds
        type: string
      message:
        description: Message is a human-readable explanation of the field violation.
        example: must be between 60 and 900
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: sharmila.dayalan@bcbpsolutions.com
    name: BCBP Solutions
  description: |-
    Organizational layer of the IAM subsystem — owns tenants, departments, memberships, tenant-level roles, delegations, invitations, ACL overlays, and group→role mappings.

    **Tenant isolation.** Every resource is scoped by `x-tenant-id`; cross-tenant reads are blocked by row-level security (RLS-1..RLS-6). The hot path `GET /internal/users/:id/memberships` is the enrichment source for AuthZ (SLO 15 ms cache-hit / 30 ms cache-miss).

    **Route prefixes.**  `/api/v1/*` — authenticated tenant users (JWT via gateway).  `/api/v1/internal/*` — in-mesh services (mTLS, `iam-system` role).  `/api/v1/operator/*` — platform operators (`platform_operator` role).

    Every mutation echoes `record_version` for optimistic-locking round-trip (CONC-4).
  license:
    name: Proprietary
  title: IAM Org & Membership API
  version: "1.0"
paths:
...
  /tenants/{id}/members:
    get:
      description: Keyset paginate on idx_tm_tenant_created. Response includes `next_cursor`
        when more pages exist.
      parameters:
      - description: Tenant UUID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      - description: Page size (1..200, default 50)
        in: query
        name: limit
        type: integer
      - description: Base64-encoded opaque cursor from previous response
        in: query
        name: cursor
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/http.MembershipListResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/http.ErrorResponse'
      security:
      - UserID: []
      - TenantID: []
      summary: P-4 — List members (cursor-paginated)
      tags:
      - members
    post:
      consumes:
      - application/json
      description: 'AUTH-2 tenant_admin/owner. SEAT-1 transactional cap: `active +
        pending <= licensed_seats`. Two-step invite→accept. Returns 409 seat_limit_reached
        at/above cap; 409 invitation_already_exists on duplicate; 429 reinvite_too_soon
        (PI-11) or invite_rate_limited (PI-12).'
      parameters:
      - description: Tenant UUID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      - description: Invite payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/http.InvitationCreateRequest'
      produces:
      - application/json
      responses:
        "202":
          description: Pending row created + RP CreateInvitedUser dispatched
          schema:
            $ref: '#/definitions/http.InvitationResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "409":
          description: seat_limit_reached OR invitation_already_exists
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "429":
          description: reinvite_too_soon OR invite_rate_limited
          schema:
            $ref: '#/definitions/http.ErrorResponse'
      security:
      - UserID: []
      - TenantID: []
      summary: P-6 — Invite user (staged pending_invitation)
      tags:
      - invitations
  /tenants/{id}/members/{user_id}:
    get:
      description: Returns the tenant_memberships row joined with tenant_roles and
        dept_memberships views.
      parameters:
      - description: Tenant UUID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      - description: User UUID
        format: uuid
        in: path
        name: user_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/http.MemberItemResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/http.ErrorResponse'
      security:
      - UserID: []
      - TenantID: []
      summary: P-5 — Read a single member (with roles + departments)
      tags:
      - members
    patch:
      consumes:
      - application/json
      description: 'AUTH-2. On suspend: AUTH-8 best-effort RP RevokeUserSessions (fail-open).
        Applies optimistic locking (CONC-4).'
      parameters:
      - description: Tenant UUID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      - description: User UUID
        format: uuid
        in: path
        name: user_id
        required: true
        type: string
      - description: Status payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/http.MembershipPatchRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/http.MembershipPatchResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "409":
          description: record_version mismatch (CONC-4)
          schema:
            $ref: '#/definitions/http.ErrorResponse'
      security:
      - UserID: []
      - TenantID: []
      summary: P-7 — Suspend/reactivate member (status only)
      tags:
      - members
  /tenants/{id}/members/{user_id}/roles:
    put:
      consumes:
      - application/json
      description: 'AUTH-2. §16 A14: one TenantRoleGranted/TenantRoleRevoked event
        per role delta. TM-8 last-owner guard → 422 last_owner_removal. `member` is
        barred (TR-7) — returns 400 invalid_role.'
      parameters:
      - description: Tenant UUID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      - description: User UUID
        format: uuid
        in: path
        name: user_id
        required: true
        type: string
      - description: Desired roles
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/http.RolesPutRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/http.RolesReconcileResponse'
        "400":
          description: invalid_role (TR-7)
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "422":
          description: last_owner_removal
          schema:
            $ref: '#/definitions/http.ErrorResponse'
      security:
      - UserID: []
      - TenantID: []
      summary: P-28 — Full-replacement multi-role reconcile
      tags:
      - members
...
  /tenants/{id}/seat-usage:
    get:
      description: AUTH-2. Returns `{active_users, pending_invitations, licensed_seats,
        over_cap, overage_since, grace_ends_at}` (SEAT-1..5).
      parameters:
      - description: Tenant UUID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/http.SeatUsageResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/http.ErrorResponse'
      security:
      - UserID: []
      - TenantID: []
      summary: P-27 — Seat-usage projection
      tags:
      - members
...
securityDefinitions:
  TenantID:
    description: Tenant UUID injected by the API gateway. Establishes the RLS scope
      for the request (RLS-6).
    in: header
    name: x-tenant-id
    type: apiKey
  TenantRoles:
    description: Comma-separated tenant-role list injected by the API gateway. Operator
      routes require `platform_operator`; internal routes require `iam-system`.
    in: header
    name: x-tenant-roles
    type: apiKey
  UserID:
    description: Authenticated user UUID injected by the API gateway. Use `iam-system`
      for in-mesh service calls hitting `/api/v1/internal/*`.
    in: header
    name: x-user-id
    type: apiKey
swagger: "2.0"